+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## master (unreleased)
 
+## 0.1.1 (2020-06-24)
+
+* RSpec custom matchers
+* RSpec helpers
+
 ## 0.1.0 (2020-06-05)
 
 * Initial release